Logout confirmation box, toast colors

+++ frontend-auth-module/projects/sl-auth-lib/src/lib/services/auth.service.ts
@@ -4,6 +4,9 @@ import { catchError, of } from 'rxjs';
 import { ApiBaseService } from './base-api/api-base.service';
 import { EndpointService } from './endpoint/endpoint.service';
 import { ToastService } from './toast/toast.service';
+import { MatDialog } from '@angular/material/dialog';
+import { ModelComponent } from '../components/shared/component/model/model.component';
+
 @Injectable({
   providedIn: 'root'
 })
@@ -14,6 +17,7 @@ export class AuthService {
   router: Router;
   configData: any;
   toastService: ToastService;
+  dialog: MatDialog;
 
   constructor() {
     this.baseApiService = inject(ApiBaseService);
@@ -21,6 +25,7 @@ export class AuthService {
     this.router = inject(Router);
     this.fetchConfigData();
     this.toastService = inject(ToastService);
+    this.dialog = inject(MatDialog);
   }
 
   async fetchConfigData() {
@@ -42,33 +47,39 @@ export class AuthService {
     return false;
   }
 
-  logout() {
-    this.loggedIn = false;
-    const payload = {
-      refresh_token: localStorage?.getItem('refToken')
-    }
+  async logout() {
+    const dialogRef = this.dialog.open(ModelComponent);
 
-    this.baseApiService
-      .post(
-        this.configData?.baseUrl,
-        this.configData?.logoutApiPath,
-        payload
-      ).pipe(
-        catchError((error) => {
-          this.toastService.showToast(error?.error?.message || `An error occurred during logout`, 'error', 3000, 'top', 'end');
-          throw error
-        })
-      )
-      .subscribe(
-        (res: any) => {
-          if (res?.responseCode === "OK") {
-            localStorage.clear();
-            this.router.navigate(['/landing']);
-          } else {
-            this.toastService.showToast(res?.message || `Logout unsuccessful`, 'error', 3000, 'top', 'end');
-          }
-        }
-      );
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.loggedIn = false;
+        const payload = {
+          refresh_token: localStorage?.getItem('refToken')
+        };
+
+        this.baseApiService
+          .post(
+            this.configData?.baseUrl,
+            this.configData?.logoutApiPath,
+            payload
+          ).pipe(
+            catchError((error) => {
+              this.toastService.showToast(error?.error?.message || `An error occurred during logout`, 'error', 3000, 'top', 'end');
+              throw error;
+            })
+          )
+          .subscribe(
+            (res: any) => {
+              if (res?.responseCode === "OK") {
+                localStorage.clear();
+                this.router.navigate(['/landing']);
+              } else {
+                this.toastService.showToast(res?.message || `Logout unsuccessful`, 'error', 3000, 'top', 'end');
+              }
+            }
+          );
+      }
+    });
   }
 
   isLoggedIn(): boolean {
